requirements.txt
amqp==2.3.2
bcrypt==3.1.4
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Django==2.1.1
django-cors-headers==2.4.0
django-filter==2.0.0
django-redis==4.9.0
djangorestframework==3.8.2
djangorestframework-filters==0.10.2
djrill==2.1.0
drf-extensions==0.4.0
factory-boy==2.11.1
Faker==0.9.1
idna==2.7
kombu==4.2.1
mock==2.0.0
pbr==4.2.0
psycopg2-binary==2.7.5
pycparser==2.19
python-dateutil==2.7.3
pytz==2018.5
raven==6.9.0
redis==2.10.6
requests==2.19.1
six==1.11.0
text-unidecode==1.2
urllib3==1.23
uWSGI==2.0.17.1
vine==1.1.4